add namespace to helper functions
